% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Separate_categoric.R
\name{Separate_categoric}
\alias{Separate_categoric}
\title{Hypothesis testing - Step 1}
\usage{
Separate_categoric(data)
}
\arguments{
\item{data}{}
}
\description{
Separate the categorical variables
}
\examples{
Separate_categoric(cars)

}
