% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Separate_numeric.R
\name{Separate_numeric}
\alias{Separate_numeric}
\title{Hypothesis testing - Step 2}
\usage{
Separate_numeric(data)
}
\arguments{
\item{data}{}
}
\description{
Separate the numeric variables
}
\examples{
Separate_numeric(cars)

}
